requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
coinbase-advanced-py>=1.0.0
websockets>=12.0
aiohttp>=3.9.0
loguru>=0.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
